qos: QOS fails with 'Attribute is not defined: QOS_POLICY'
Motivation: Find out where in the code we throw the above error and resolve it.

Modification: Removal of redundant qos policy logic.

Result: Instead of checking if the qos_policy attribute is defined AND null, we will check if it is present using one of the predefined methods.
If not, set it to null and resume previous logic.

Acked-by: Tigran Mkrtchyan
Target: master
Require-book: no
Require-notes: no